fix(loadgen): 线程数=并发槽,窗口外机器人记 skipped
一机器人一线程在万号档把上万线程压进调度器(本地实测 M3 全程烧掉
~500 CPU 秒系统态),macOS 还有每进程线程数上限:线程创建失败抛
system_error,unwind 穿过仍 joinable 的 vector<thread> 直接裸
terminate(CI 背栈证实)。改为线程数 = 并发槽数,机器人从原子游标
依次领号;ThreadJoiner 保证异常路径也先 join 再传播,真实异常类型
可到 gtest 手里。

窗口关闭时没起跑的机器人记 skipped 单列,不再伪装成拨号超时掩盖
真实吞吐(CI round3 实测 verify p99 85ms 全是已过期的截止残片)。
顺带修 alpn_wire.size() 的 64→32 位截断告警。

本地 M3:58s → 25s,系统态 500 CPU 秒 → 4.4s;utility 限频档
39.6s 仍绿。

+++ tools/loadgen/src/tls_wire.hpp
@@ -188,7 +188,7 @@ struct Stream final {
         SSL_set_alpn_protos(
             out.ssl.get(),
             reinterpret_cast<const unsigned char*>(alpn_wire.data()),
-            alpn_wire.size()) != 0) {
+            static_cast<unsigned int>(alpn_wire.size())) != 0) {
         ++dial_failures.ssl_setup;
         return false;
     }
